+++ Assets/Graphics/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae6f8d2d341200a4f96e646f82f8cd78
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
