+++ Assets/Graphics/Fonts/KiwiMaru_OFL.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 448e95f1a6d8dc648ace8ff68099d21a
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
