+++ Assets/Graphics/Fonts/NewTegomin_OFL.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6d8a8fdfa89348c4e9aa7cd52bf8842d
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
