+++ Assets/Prefabs/UI/Slider.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3696132df5aa73a4b98a385bee9d563e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
