--- Assets/Scripts/Lobby/HostLobbyManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 16ea6f9d0524074448729a9dca0d4bfc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
