--- Assets/Scripts/Lobby/JoinLobbyManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bcb540db64388ad46b57c0543cb77575
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
