+++ Assets/Scripts/Lobby/SteamLobby.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46dd845e7b37dc44b87bc9e7484a123d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
