+++ Assets/Scripts/OLD_LOCAL_CONNECT.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b32802adebe69b468d6efaae9a2b4cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
